# Rewrite portable implementation for slight perf win.