~3% speed improvement on the M1:

```
Xoodoo permutation      time:   [71.380 ns 71.439 ns 71.498 ns]                               
                        change: [-3.5832% -3.4892% -3.3789%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

Xoodyak hash            time:   [1.2899 us 1.2905 us 1.2913 us]                          
                        change: [-1.7372% -1.4566% -1.1911%] (p = 0.00 < 0.05)
                        Performance has improved.

Xoodyak keyed           time:   [550.43 ns 550.63 ns 550.83 ns]                           
                        change: [-1.8320% -1.7599% -1.6869%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
```

This translates into similarly mild improvements in real-world usage. A test app which hashes a 100MiB file:

```
Benchmark 1: control
  Time (mean ± σ):     494.6 ms ±   0.5 ms    [User: 495.4 ms, System: 33.0 ms]
  Range (min … max):   493.5 ms … 495.0 ms    10 runs
 
Benchmark 2: experimental
  Time (mean ± σ):     483.4 ms ±   0.7 ms    [User: 484.4 ms, System: 32.2 ms]
  Range (min … max):   482.8 ms … 485.2 ms    10 runs
 
Summary
  'experimental' ran
    1.02 ± 0.00 times faster than ‘control'
```

That said, it uses a little more memory and may well be slower on other processors/architectures. I’d totally understand if you’d prefer the existing implementation.